+++ baldaquin/pkt.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2022--2024 the baldaquin team.
+# Copyright (C) 2022--2025 the baldaquin team.
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -16,11 +16,15 @@
 """Binary data packet utilities.
 """
 
+from __future__ import annotations
+
 from abc import ABC, abstractmethod
+from contextlib import contextmanager
 from dataclasses import dataclass
 from enum import Enum
 import struct
 
+from baldaquin import logger, __version__
 from baldaquin.timeline import Timeline
 
 
@@ -77,16 +81,13 @@ class AbstractPacket(ABC):
     """Abstract base class for binary packets.
     """
 
-    COMMENT_PREFIX = '# '
-    TEXT_SEPARATOR = ', '
-
     def __post_init__(self) -> None:
         """Hook for post-initialization.
         """
 
     @property
     @abstractmethod
-    def payload(self) -> bytes:
+    def data(self) -> bytes:
         """Return the packet binary data.
         """
 
@@ -117,14 +118,86 @@ class AbstractPacket(ABC):
         """Unpack the binary data into the corresponding field values.
         """
 
-    @classmethod
-    def text_header(cls) -> str:
+    def _format_attributes(self, attrs: tuple[str], fmts: tuple[str] = None) -> tuple[str]:
+        """Helper function to join a given set of class attributes in a properly
+        formatted string.
+
+        This is used, most notably, in the :meth:`_repr() <baldaquin.pkt.AbstractPacket._repr>`
+        hook below, which in turn is used in the various ``__repr__()`` and/or ``__str__``
+        implementations, and in the :meth:`to_text() <baldaquin.pkt.AbstractPacket.to_text>`
+        implementations in sub-classes.
+
+        Arguments
+        ---------
+        attrs : tuple
+            The names of the class attributes we want to include in the representation.
+
+        fmts : tuple, optional
+            If present determines the formatting of the given attributes.
+        """
+        vals = (getattr(self, attr) for attr in attrs)
+        if fmts is None:
+            fmts = ('%s' for _ in attrs)
+        return tuple(fmt % val for val, fmt in zip(vals, fmts))
+
+    def _text(self, attrs: tuple[str], fmts: tuple[str], separator: str) -> str:
+        """Helper function for text formatting.
+
+        Note the output includes a trailing endline.
+
+        Arguments
+        ---------
+        attrs : tuple
+            The names of the class attributes we want to include in the representation.
+
+        fmts : tuple,
+            Determines the formatting of the given attributes.
+
+        separator : str
+            The separator between different fields.
+        """
+        vals = self._format_attributes(attrs, fmts)
+        return f'{separator.join(vals)}\n'
+
+    def _repr(self, attrs: tuple[str], fmts: tuple[str] = None) -> str:
+        """Helper function to provide sensible string formatting for the packets.
+
+        The basic idea is that concrete classes would use this to implement their
+        `__repr__()` and/or `__str__()` special dunder methods.
+
+        Arguments
+        ---------
+        attrs : tuple
+            The names of the class attributes we want to include in the representation.
+
+        fmts : tuple, optional
+            If present determines the formatting of the given attributes.
+        """
+        vals = self._format_attributes(attrs, fmts)
+        info = ', '.join([f'{attr}={val}' for attr, val in zip(attrs, vals)])
+        return f'{self.__class__.__name__}({info})'
+
+    @staticmethod
+    def text_header(prefix: str, creator: str = None) -> str:
         """Hook that subclasses can overload to provide a sensible header for an
         output text file.
-        """
-        return f'{cls.COMMENT_PREFIX}Created on {Timeline().latch()}'
 
-    def to_text(self) -> str:
+        Arguments
+        ---------
+        prefix : str
+            The prefix to be prepended to each line to signal that that line is
+            a comment and contains no data.
+
+        creator : str, optional
+            An optional string indicating the application that created the file.
+        """
+        header = f'{prefix}Created on {Timeline().latch()}\n' \
+                 f'{prefix}baldaquin version: {__version__}\n'
+        if creator is not None:
+            header = f'{header}{prefix}Creator: {creator}\n'
+        return header
+
+    def to_text(self, separator: str) -> str:
         """Hook that subclasses can overload to provide a text representation of
         the buffer to be written in an output text file.
         """
@@ -146,14 +219,28 @@ class FieldMismatchError(RuntimeError):
 def _class_annotations(cls) -> dict:
     """Small convienience function to retrieve the class annotations.
 
-    This is needed because in Python 3.7 cls.__annotations__ is not defined
-    when a class has no annotations, while in subsequent Python versions an empty
-    dictionary is returned, instead.
+    Note that, in order to support inheritance of @packetclasses, we do iterate
+    over all the ancestors of the class at hand, starting from ``AbstractPacket``,
+    and collect all the annotations along the way. The iteration is in reverse
+    order, so that the final order of annotations is what one would expect.
+
+    The try/except clause is needed because in Python 3.7 cls.__annotations__ is
+    not defined when a class has no annotations, while in subsequent Python
+    versions an empty dictionary is returned, instead.
     """
-    try:
-        return cls.__annotations__
-    except AttributeError:
-        return {}
+    # The typical mro, here, is object -> ABC -> AbstractPacket -> ...
+    # Although object and ABC seem to have no annotations, and might very well
+    # never have one, we only start the annotation-collection loop from
+    # AbstractPacket, so that the situation is under our full control (cross your)
+    # fingers.
+    ancestors = cls.__mro__[:cls.__mro__.index(AbstractPacket)]
+    annotations = {}
+    for _cls in reversed(ancestors):
+        try:
+            annotations.update(_cls.__annotations__)
+        except AttributeError:
+            pass
+    return annotations
 
 
 def _check_format_characters(cls: type) -> None:
@@ -184,9 +271,9 @@ def packetclass(cls: type) -> type:
     cls._format = f'{cls.layout.value}{"".join(char.value for char in annotations.values())}'
     cls.size = struct.calcsize(cls._format)
     # And here is a list of attributes we want to be frozen.
-    cls.__frozenattrs__ = ('_fields', '_format', 'size', '_payload') + cls._fields
+    cls.__frozenattrs__ = ('_fields', '_format', 'size', '_data') + cls._fields
 
-    def _init(self, *args, payload: bytes = None):
+    def _init(self, *args, data: bytes = None):
         # Make sure we have the correct number of arguments---they should match
         # the class annotations.
         if len(args) != len(cls._fields):
@@ -200,9 +287,9 @@ def packetclass(cls: type) -> type:
             if expected is not None and expected != value:
                 raise FieldMismatchError(cls, field, expected, value)
             object.__setattr__(self, field, value)
-        if payload is None:
-            payload = self.pack()
-        object.__setattr__(self, '_payload', payload)
+        if data is None:
+            data = self.pack()
+        object.__setattr__(self, '_data', data)
         # Make sure the post-initialization is correctly performed.
         self.__post_init__()
 
@@ -222,11 +309,11 @@ class FixedSizePacketBase(AbstractPacket):
     _format = None
     size = 0
     __frozenattrs__ = None
-    _payload = None
+    _data = None
 
     @property
-    def payload(self) -> bytes:
-        return self._payload
+    def data(self) -> bytes:
+        return self._data
 
     @property
     def fields(self) -> tuple:
@@ -243,7 +330,7 @@ class FixedSizePacketBase(AbstractPacket):
 
     @classmethod
     def unpack(cls, data: bytes) -> AbstractPacket:
-        return cls(*struct.unpack(cls._format, data), payload=data)
+        return cls(*struct.unpack(cls._format, data), data=data)
 
     def __setattr__(self, key, value) -> None:
         """Overloaded method to make class instances frozen.
@@ -252,24 +339,80 @@ class FixedSizePacketBase(AbstractPacket):
             raise AttributeError(f'Cannot modify {self.__class__.__name__}.{key}')
         object.__setattr__(self, key, value)
 
+    def __repr__(self):
+        """String formatting.
+
+        Note that this provides the low-level string formatting, according to the
+        typical Python conventions. Subclasses are welcome to provide more
+        succinct and human-readable `__str__()` special dunder implementations, but
+        are encouraged not to reimplement this special method, as it inclues all
+        the nitty-gritty details of the packet, which might be useful when debugging.
+        """
+        return self._repr(self._fields + ('data', '_format'))
+
     def __str__(self):
         """String formatting.
+
+        This is a slightly more concise test representation of the packet. Again,
+        subclasses can reimplement this at their leisure.
         """
-        attrs = self._fields + ('payload', '_format')
-        info = ', '.join([f'{attr}={getattr(self, attr)}' for attr in attrs])
-        return f'{self.__class__.__name__}({info})'
+        return self._repr(self._fields)
 
     @classmethod
-    def text_header(cls) -> str:
+    def text_header(cls, prefix: str, creator: str = None) -> str:
         """Overloaded method.
         """
-        return f'{super().text_header()}\n' \
-               f'{cls.COMMENT_PREFIX}{cls.TEXT_SEPARATOR.join(cls._fields)}\n'
+        return f'{AbstractPacket.text_header(prefix, creator)}' \
+               f'{prefix}{", ".join(cls._fields)}\n'
 
-    def to_text(self) -> str:
+    def to_text(self, separator: str) -> str:
         """Overloaded method.
         """
-        return f'{self.TEXT_SEPARATOR.join([str(item) for item in self])}\n'
+        return f'{separator.join([str(item) for item in self])}\n'
+
+
+class PacketFile:
+
+    """Class describing a binary file containing packets.
+    """
+
+    def __init__(self, packet_class: type) -> None:
+        """Constructor.
+        """
+        self._packet_class = packet_class
+        self._input_file = None
+
+    @contextmanager
+    def open(self, file_path: str):
+        """Open the file.
+        """
+        logger.debug(f'Opening input packet file {file_path}...')
+        with open(file_path, 'rb') as input_file:
+            self._input_file = input_file
+            yield self
+            self._input_file = None
+        logger.debug(f'Input file {file_path} closed.')
+
+    def __iter__(self) -> 'PacketFile':
+        """Return the iterator object (self).
+        """
+        return self
+
+    def __next__(self) -> FixedSizePacketBase:
+        """Read the next packet in the buffer.
+        """
+        data = self._input_file.read(self._packet_class.size)
+        if not data:
+            raise StopIteration
+        return self._packet_class.unpack(data)
+
+    def read_all(self) -> tuple[FixedSizePacketBase]:
+        """Read in memory all the packets in the file.
+
+        This is meant to support postprocessing applications where one needs
+        all the packets in memory at the same time. Use it `cum grano salis`.
+        """
+        return tuple(packet for packet in self)
 
 
 @dataclass