+++ baldaquin/plasduino/protocol.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2024 the baldaquin team.
+# Copyright (C) 2024--2025 the baldaquin team.
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -16,9 +16,13 @@
 """Basic definition of the plasduino communication protocol.
 """
 
-from enum import Enum
+from enum import Enum, IntEnum
 
-from baldaquin.pkt import packetclass, FixedSizePacketBase, Format, Layout
+from baldaquin.pkt import packetclass, AbstractPacket, FixedSizePacketBase, Format, Layout
+
+
+COMMENT_PREFIX = '# '
+TEXT_SEPARATOR = ', '
 
 
 class Marker(Enum):
@@ -64,35 +68,49 @@ class OpCode(Enum):
     OP_CODE_TOGGLE_DIGITAL_PIN = 0x0D
 
 
-@packetclass
-class DigitalTransition(FixedSizePacketBase):
+class InterruptMode(IntEnum):
 
-    """A plasduino digital transition is a 6-bit binary array containing:
-
-    * byte(s) 0  : the array header (``Marker.DIGITAL_TRANSITION_HEADER.value``);
-    * byte(s) 1  : the transition information (pin number and edge type);
-    * byte(s) 2-5: the timestamp of the readout from micros().
+    """Definition of the interrupt modes.
     """
 
-    layout = Layout.BIG_ENDIAN
-    header: Format.UNSIGNED_CHAR = Marker.DIGITAL_TRANSITION_HEADER.value
-    info: Format.UNSIGNED_CHAR
-    microseconds: Format.UNSIGNED_LONG
-
-    def __post_init__(self) -> None:
-        """Post initialization.
-        """
-        # Note the _info field is packing into a single byte the edge type
-        # (the MSB) and the pin number.
-        self.pin_number = self.info & 0x7F
-        self.edge = (self.info >> 7) & 0x1
-        self.seconds = 1.e-6 * self.seconds
+    DISABLED = 0
+    CHANGE = 1
+    FALLING = 2
+    RISING = 3
 
 
 @packetclass
-class AnalogReadout(FixedSizePacketBase):
+class PlasduinoPacketBase(FixedSizePacketBase):
 
-    """A plasduino analog readout is a 8-bit binary array containing:
+    """Base class for the plasduino packets.
+    """
+
+    OUTPUT_HEADERS = None
+    OUTPUT_ATTRIBUTES = None
+    OUTPUT_FMTS = None
+
+    @classmethod
+    def text_header(cls, prefix: str = COMMENT_PREFIX, creator: str = None) -> str:
+        """Return the header for the output text file.
+        """
+        headers = ', '.join(map(str, cls.OUTPUT_HEADERS))
+        return f'{AbstractPacket.text_header(prefix, creator)}{prefix}{headers}\n'
+
+    def to_text(self, separator: str = TEXT_SEPARATOR) -> str:
+        """Convert a readout to text for use in a custom sink.
+        """
+        return self._text(self.OUTPUT_ATTRIBUTES, self.OUTPUT_FMTS, separator)
+
+    def __str__(self):
+        """String formatting.
+        """
+        return self._repr(self.OUTPUT_ATTRIBUTES, self.OUTPUT_FMTS)
+
+
+@packetclass
+class AnalogReadout(PlasduinoPacketBase):
+
+    """A plasduino analog readout is a 8-byte binary array containing:
 
     * byte(s) 0  : the array header (``Marker.ANALOG_READOUT_HEADER.value``);
     * byte(s) 1  : the analog pin number;
@@ -106,7 +124,40 @@ class AnalogReadout(FixedSizePacketBase):
     milliseconds: Format.UNSIGNED_LONG
     adc_value: Format.UNSIGNED_SHORT
 
+    OUTPUT_HEADERS = ('Pin number', 'Time [s]', 'ADC counts')
+    OUTPUT_ATTRIBUTES = ('pin_number', 'seconds', 'adc_value')
+    OUTPUT_FMTS = ('%d', '%.3f', '%d')
+
     def __post_init__(self) -> None:
         """Post initialization.
         """
         self.seconds = 1.e-3 * self.milliseconds
+
+
+@packetclass
+class DigitalTransition(PlasduinoPacketBase):
+
+    """A plasduino digital transition is a 6-byte binary array containing:
+
+    * byte(s) 0  : the array header (``Marker.DIGITAL_TRANSITION_HEADER.value``);
+    * byte(s) 1  : the transition information (pin number and edge type);
+    * byte(s) 2-5: the timestamp of the readout from micros().
+    """
+
+    layout = Layout.BIG_ENDIAN
+    header: Format.UNSIGNED_CHAR = Marker.DIGITAL_TRANSITION_HEADER.value
+    info: Format.UNSIGNED_CHAR
+    microseconds: Format.UNSIGNED_LONG
+
+    OUTPUT_HEADERS = ('Pin number', 'Edge type', 'Time [s]')
+    OUTPUT_ATTRIBUTES = ('pin_number', 'edge', 'seconds')
+    OUTPUT_FMTS = ('%d', '%d', '%.6f')
+
+    def __post_init__(self) -> None:
+        """Post initialization.
+        """
+        # Note the _info field is packing into a single byte the edge type
+        # (the MSB) and the pin number.
+        self.pin_number = self.info & 0x7F
+        self.edge = (self.info >> 7) & 0x1
+        self.seconds = 1.e-6 * self.microseconds